moves injector to injector directory
